requirements/base.txt
Django==1.6.5
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
bpython==0.13
coverage==3.7.1
django-allauth==0.17.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-model-utils==2.0.3
djrill==1.1.1
docutils==0.12
git-remote-helpers==0.1.0
logutils==0.3.3
oauthlib==0.6.3
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
